requirement.txt
Flask==3.0.0
mysql-connector-python==8.2.0
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3